requirements.txt
avalanche-lib>=0.6.0
pandas
numpy
torch